[REF] Conform logic to the designed one

+++ images/common/spank-nccl-debug/src/snccld.c
@@ -1,7 +1,8 @@
 #include "snccld.h"
 #include "snccld_args.h"
-#include "snccld_mkdir.h"
 #include "snccld_state.h"
+#include "snccld_util_antidupl.h"
+#include "snccld_util_dir.h"
 
 #include <ctype.h>
 #include <fcntl.h>
@@ -105,6 +106,115 @@ int slurm_spank_init(spank_t spank, int argc, char **argv) {
     }
 }
 
+static spank_err_t snccld_prepare_directories() {
+    slurm_spank_log(
+        "%s: Prepare directories: '%s', '%s'.",
+        SNCCLD_LOG_PREFIX,
+        snccld_config.out_dir,
+        SNCCLD_DEFAULT_DIR
+    );
+
+    if (snccld_mkdir_p(snccld_config.out_dir, SNCCLD_DEFAULT_MODE) ==
+        ESPANK_ERROR) {
+        slurm_error(
+            "%s: Cannot create directory %s: %m",
+            SNCCLD_LOG_PREFIX,
+            snccld_config.out_dir
+        );
+        return ESPANK_ERROR;
+    }
+
+    if (snccld_mkdir_p(SNCCLD_DEFAULT_DIR, SNCCLD_DEFAULT_MODE) ==
+        ESPANK_ERROR) {
+        slurm_error(
+            "%s: Cannot create directory '%s': %m",
+            SNCCLD_LOG_PREFIX,
+            SNCCLD_DEFAULT_DIR
+        );
+        return ESPANK_ERROR;
+    }
+
+    return ESPANK_SUCCESS;
+}
+
+static void snccld_run_named_pipe_reading_process(
+    snccld_state_t *state, const bool user_set_debug_file, char *user_debug_file
+) {
+    // Build shell command: <SHELL> -c '<COMMAND>'
+    char *sh_argv[3];
+    int   sh_idx = 0;
+
+    // Choose shell.
+    {
+        sh_argv[sh_idx++] =
+            (access("/bin/bash", X_OK) == 0) ? "/bin/bash" : "/bin/sh";
+        sh_argv[sh_idx++] = "-c";
+    }
+
+    // Build tee command: <STDBUF> -oL <TEE> -a [TARGETS]
+    char *tee_argv[11];
+    int   tee_idx = 0;
+
+    // Use line buffering instead of tee's internal buffer.
+    tee_argv[tee_idx++] = "/usr/bin/stdbuf";
+    tee_argv[tee_idx++] = "-oL";
+
+    // Run tee to append to files.
+    tee_argv[tee_idx++] = "/usr/bin/tee";
+    tee_argv[tee_idx++] = "-a";
+
+    // Construct a list of unique targets.
+    char *output_targets[2];
+    {
+        size_t target_count = 0;
+
+        // Include user-specified debug file.
+        if (user_set_debug_file) {
+            output_targets[target_count++] = user_debug_file;
+        }
+
+        // Include out file.
+        if (snccld_config.out_file) {
+            output_targets[target_count++] = state->log_path;
+        }
+
+        size_t unique_output_targets =
+            snccld_remove_string_duplicates(output_targets, target_count);
+        for (size_t i = 0; i < unique_output_targets; ++i) {
+            tee_argv[tee_idx++] = output_targets[i];
+        }
+    }
+
+    // Take input from named pipe.
+    tee_argv[tee_idx++] = "<";
+    tee_argv[tee_idx++] = state->fifo_path;
+
+    // Suppress stdout.
+    if (!snccld_config.out_stdout) {
+        tee_argv[tee_idx++] = ">";
+        tee_argv[tee_idx++] = "/dev/null";
+    }
+
+    {
+        size_t buf_size    = PATH_MAX * 3;
+        char  *tee_command = malloc(buf_size);
+        size_t offset      = 0;
+        for (size_t i = 0; i < tee_idx; ++i) {
+            offset += snprintf(
+                tee_command + offset, buf_size - offset, "%s ", tee_argv[i]
+            );
+        }
+        sh_argv[sh_idx++] = tee_command;
+    }
+    sh_argv[sh_idx] = NULL;
+
+    slurm_spank_log("Running: %s -c '%s'", sh_argv[0], sh_argv[2]);
+
+    execvp(sh_argv[0], sh_argv);
+
+    _exit(EXIT_FAILURE);
+}
+
 int slurm_spank_user_init(spank_t spank, int argc, char **argv) {
     log_context("user_init", spank);
 
@@ -137,112 +247,139 @@ int slurm_spank_user_init(spank_t spank, int argc, char **argv) {
         free(key);
         return ESPANK_SUCCESS;
     }
-
     snccld_state_t *state = snccld_state_new();
 
-    char debug_val[16]  = "";
-    int  user_set_debug = 0;
-    if (spank_getenv(
-            spank, SNCCLD_NCCL_ENV_DEBUG, debug_val, sizeof(debug_val)
-        ) == ESPANK_SUCCESS) {
-        user_set_debug = 1;
-    }
+    // Set forced debug level.
+    slurm_spank_log(
+        "%s: Setting %s=%s",
+        SNCCLD_LOG_PREFIX,
+        SNCCLD_NCCL_ENV_DEBUG,
+        snccld_config.log_level
+    );
+    spank_setenv(spank, SNCCLD_NCCL_ENV_DEBUG, snccld_config.log_level, 1);
 
-    if (snccld_mkdir_p(SNCCLD_DEFAULT_DIR, SNCCLD_DEFAULT_FIFO_MODE) ==
-        ESPANK_ERROR) {
-        slurm_error(
-            SNCCLD_LOG_PREFIX "Cannot create directory '%s': %m",
-            SNCCLD_DEFAULT_DIR
+    // Check if user set debug file.
+    static char user_debug_file[PATH_MAX] = "";
+    bool        user_set_debug_file =
+        (spank_getenv(
+             spank,
+             SNCCLD_NCCL_ENV_DEBUG_FILE,
+             user_debug_file,
+             sizeof(user_debug_file)
+         ) == ESPANK_SUCCESS);
+    slurm_spank_log(
+        "%s: user_set_debug_file =%u", SNCCLD_LOG_PREFIX, user_set_debug_file
+    );
+
+    // Neither outfile nor stdout requested nor user set debug file -> noop
+    if (!snccld_config.out_file && !snccld_config.out_stdout &&
+        !user_set_debug_file) {
+        slurm_spank_log(
+            "%s: Neither out file nor stdout requested nor user set debug "
+            "file. "
+            "Skipping.",
+            SNCCLD_LOG_PREFIX
         );
         free(key);
         free(state);
-        return ESPANK_ERROR;
+        return ESPANK_SUCCESS;
     }
+
+    // Ensure necessary directories exist.
+    if (snccld_config.out_file || snccld_config.out_stdout ||
+        user_set_debug_file) {
+        if (snccld_prepare_directories() != ESPANK_SUCCESS) {
+            free(key);
+            free(state);
+            return ESPANK_ERROR;
+        }
+    }
+
+    if (snccld_config.out_file) {
+        snprintf(
+            state->log_path,
+            sizeof(state->log_path),
+            SNCCLD_TEMPLATE_FILE_NAME,
+            snccld_config.out_dir,
+            key->job_id,
+            key->step_id,
+            "out"
+        );
+    }
+
+    // Only out file needed -> set NCCL_DEBUG_FILE.
+    if (snccld_config.out_file && !snccld_config.out_stdout &&
+        !user_set_debug_file) {
+        slurm_spank_log("%s: Only out file needed.", SNCCLD_LOG_PREFIX);
+        slurm_spank_log(
+            "%s: Setting %s=%s",
+            SNCCLD_LOG_PREFIX,
+            SNCCLD_NCCL_ENV_DEBUG_FILE,
+            state->log_path
+        );
+        spank_setenv(spank, SNCCLD_NCCL_ENV_DEBUG_FILE, state->log_path, 1);
+
+        free(key);
+        free(state);
+
+        return ESPANK_SUCCESS;
+    }
+
+    // If we're here, named pipe has to be constructed.
+
+    slurm_spank_log("%s: Named pipe has to be constructed.", SNCCLD_LOG_PREFIX);
     snprintf(
         state->fifo_path,
         sizeof(state->fifo_path),
-        "%s/%u_%u.fifo",
+        SNCCLD_TEMPLATE_FILE_NAME,
         SNCCLD_DEFAULT_DIR,
         key->job_id,
-        key->step_id
+        key->step_id,
+        "fifo"
     );
-
-    if (snccld_mkdir_p(snccld_config.out_dir, SNCCLD_DEFAULT_FIFO_MODE) ==
-        ESPANK_ERROR) {
+    if (mkfifo(state->fifo_path, SNCCLD_DEFAULT_MODE) != 0 && errno != EEXIST) {
         slurm_error(
-            SNCCLD_LOG_PREFIX "Cannot create directory '%s': %m",
-            snccld_config.out_dir
+            "%s: Cannot create named pipe '%s': %m",
+            SNCCLD_LOG_PREFIX,
+            state->fifo_path
         );
+
+        free(key);
+        free(state);
+
+        return ESPANK_ERROR;
+    }
+
+    slurm_spank_log(
+        "%s: Setting %s=%s",
+        SNCCLD_LOG_PREFIX,
+        SNCCLD_NCCL_ENV_DEBUG_FILE,
+        state->fifo_path
+    );
+    spank_setenv(spank, SNCCLD_NCCL_ENV_DEBUG_FILE, state->fifo_path, 1);
+
+    // Create separate process to distribute logs from the named pipe.
+    pid_t tee_pid = fork();
+    if (tee_pid < 0) {
+        // Forking failed.
+        slurm_error(
+            "%s: Cannot create named pipe reading process: %m",
+            SNCCLD_LOG_PREFIX
+        );
+        unlink(state->fifo_path);
         free(key);
         free(state);
         return ESPANK_ERROR;
-    }
-    snprintf(
-        state->log_path,
-        sizeof(state->log_path),
-        "%s/%u_%u.out",
-        snccld_config.out_dir,
-        key->job_id,
-        key->step_id
-    );
-
-    if (mkfifo(state->fifo_path, SNCCLD_DEFAULT_FIFO_MODE) != EXIT_SUCCESS) {
-        if (errno == EEXIST) {
-            unlink(state->fifo_path);
-            if (mkfifo(state->fifo_path, SNCCLD_DEFAULT_FIFO_MODE) !=
-                EXIT_SUCCESS) {
-                slurm_error(
-                    SNCCLD_LOG_PREFIX "Cannot create FIFO %s: %m",
-                    state->fifo_path
-                );
-                free(key);
-                free(state);
-                return ESPANK_SUCCESS;
-            }
-        } else {
-            slurm_error(
-                SNCCLD_LOG_PREFIX "Cannot create FIFO %s: %m", state->fifo_path
-            );
-            free(key);
-            free(state);
-            return ESPANK_SUCCESS;
-        }
-    }
-
-    pid_t pid = fork();
-    if (pid < 0) {
-        slurm_error(SNCCLD_LOG_PREFIX "fork() failed: %m");
-        unlink(state->fifo_path);
-        return ESPANK_SUCCESS;
-    } else if (pid == 0) {
-        int fd = open(state->fifo_path, O_RDONLY);
-        if (fd < 0) {
-            _exit(EXIT_FAILURE);
-        }
-        dup2(fd, STDIN_FILENO);
-        close(fd);
-
-        if (!user_set_debug) {
-            int devnull = open("/dev/null", O_WRONLY);
-            if (devnull >= 0) {
-                dup2(devnull, STDOUT_FILENO);
-                close(devnull);
-            }
-        }
-
-        execlp(
-            "stdbuf",
-            "stdbuf",
-            "-oL",
-            "/usr/bin/tee",
-            "-a",
-            state->log_path,
-            (char *)NULL
+    } else if (tee_pid == 0) {
+        // We're in forked process.
+        snccld_run_named_pipe_reading_process(
+            state, user_set_debug_file, user_debug_file
         );
-        _exit(EXIT_FAILURE);
+    } else {
+        // We're in main process -> tee_pid is a pid of the forked process.
+        state->tee_pid = tee_pid;
     }
 
-    state->tee_pid = pid;
     if (snccld_state_write(key, state) != ESPANK_SUCCESS) {
         free(key);
         free(state);
@@ -250,28 +387,8 @@ int slurm_spank_user_init(spank_t spank, int argc, char **argv) {
     }
 
     char *str = snccld_state_to_string(state);
-    slurm_spank_log(SNCCLD_LOG_PREFIX "state: %s", str);
+    slurm_spank_log("%s: State: \n%s", SNCCLD_LOG_PREFIX, str);
     free(str);
-
-    if (!user_set_debug) {
-        slurm_spank_log(
-            SNCCLD_LOG_PREFIX "Setting %s=%s",
-            SNCCLD_NCCL_ENV_DEBUG,
-            snccld_config.log_level
-        );
-        spank_setenv(spank, SNCCLD_NCCL_ENV_DEBUG, snccld_config.log_level, 1);
-    } else {
-        slurm_spank_log(SNCCLD_LOG_PREFIX "Skipping env var");
-    }
-
-    {
-        slurm_spank_log(
-            SNCCLD_LOG_PREFIX "Setting %s=%s",
-            SNCCLD_NCCL_ENV_DEBUG_FILE,
-            state->fifo_path
-        );
-        spank_setenv(spank, SNCCLD_NCCL_ENV_DEBUG_FILE, state->fifo_path, 1);
-    }
 
     free(key);
     free(state);
@@ -315,7 +432,6 @@ int slurm_spank_task_exit(spank_t spank, int argc, char **argv) {
     snccld_state_t *state = snccld_state_read(key);
     if (state == NULL) {
         free(key);
-        free(state);
         return ESPANK_ERROR;
     }
 
@@ -323,7 +439,13 @@ int slurm_spank_task_exit(spank_t spank, int argc, char **argv) {
     slurm_spank_log(SNCCLD_LOG_PREFIX "state: %s", str);
     free(str);
 
+    // Kill process reading from named pipe if exists.
     if (state->tee_pid > 0) {
+        slurm_spank_log(
+            "%s: Killing named pipe reading process with pid %d.",
+            SNCCLD_LOG_PREFIX,
+            state->tee_pid
+        );
         int status;
         if (waitpid(state->tee_pid, &status, WNOHANG) == 0) {
             kill(state->tee_pid, SIGKILL);
@@ -331,7 +453,16 @@ int slurm_spank_task_exit(spank_t spank, int argc, char **argv) {
         }
         state->tee_pid = -1;
     }
-    unlink(state->fifo_path);
+
+    // Remove named pipe if exists.
+    if (state->fifo_path) {
+        slurm_spank_log(
+            "%s: Removing named pipe '%s'.", SNCCLD_LOG_PREFIX, state->fifo_path
+        );
+        unlink(state->fifo_path);
+    } else {
+        slurm_spank_log("%s: No named pipe to remove.", SNCCLD_LOG_PREFIX);
+    }
 
     str = snccld_state_to_string(state);
     slurm_spank_log(SNCCLD_LOG_PREFIX "state: %s", str);
